first pass for data model refactor
