+++ src/imports/collections/exposureResult/constants.js
@@ -1,0 +1,17 @@
+let samplingApproaches = [
+        'Personal',
+        'Environmental',
+        'Biological',
+        'Other',
+        'Not-specified',
+    ],
+    exposureLevelDescriptions = [
+        'Arithmetic mean',
+        'Geometric mean',
+        'Median',
+        'Other',
+        'Not-reported',
+    ];
+
+export { samplingApproaches };
+export { exposureLevelDescriptions };
